Remove jitana::variable
jitana::variable was designed unify a register, an instance field
variable, and a static field variable. However, it is better to use
boost::variant for this purpose as it is more flexible.

Also, an exit instruction now takes a register as an argument. If the
its container method returns a non-void value, the argument is set to
register_idx::idx_result so that uses() on the exit instruction returns
vR in this case.