TraVis: pull JAR/APK instead of ODEX from the device
We now use apktool for extracting APK/JAR files.

The Dalvik VM keeps a copy of an APK/JAR/DEX file to /data/jitana/
whenever it loads the file to the memory. The following is the commit
message from my Dalvik patch (see tools/jitana-travis/jitana-patches/):

I was originally planning to modify dexopt to copy the DEX files loaded
by the virtual machine to safe locations. The advantage of this
approach is that it has zero overhead for the second time of execution
because dexopt is executed only when an optimized DEX file for a DEX
file to be loaded on the virtual machine. However, there are two
disadvantages to this approach:

- The optimized DEX files must be deleted when we first run the
  modified Dalvik to make copies for these DEX files. The optimized DEX
  files may be placed in directories other than /data/dalvik-cache. For
  example, the secondary DEX files for Facebook are placed in
  /data/data/com.facebook.katana/app_secondary_program_dex/.

- It cannot keep additional files such as AndroidManifest.xml with the
  DEX files.

It seems Dalvik_dalvik_system_DexFile_openDexFileNative() is executed
whenever a DEX file is loaded after Zygote. Unless an uncompressed DEX
file is explicitly loaded (which can happen in the command line mode),
a DEX file is loaded from a APK file or a JAR file. So I decided to
keep a copy of the enclosing APK/JAR file within this function. A DEX
file is copied only if it is loaded explicitly without a APK/JAR file.

A APK/JAR/DEX file is copied to /data/jitana only if a copy with the
same name and timestamp does not exist. The target file is named just
like the optimized DEX files are named. For example, /data/app/abc.apk
is copied to /data/jitana/data@app@abc.apk.

Note that the target directory should be created beforehand via adb
shell like this:

    adb shell 'mkdir /data/jitana; chmod 777 /data/jitana'

There are ways to make this permission more restrictive, but they
require more modifications to the virtual machine.