TraVis: use `adb shell ps` instead of `adb jdwp`
The JDWP number is a pid of the target app. We now use the device's `ps`
command to extract the pid using the process name. This should be more
reliable and quicker than the buggy `adb jdwp` command.

This change also allows the process name (:suffix) to be specified in
the component name (see the Facebook examples).